refactor builders.postCommentIdentifier, numericalCompare
