+++ lib/core/options/index.js
@@ -8,3 +8,5 @@ export {
 export * as stage from './stage';
 
 export * as table from './table';
+
+export * as builders from './builders';
