+++ lib/utils/index.js
@@ -116,6 +116,7 @@ export {
 
 export {
 	firstValid,
+	numericalCompare,
 } from './value';
 
 export {
